Move MacroList to model
