+++ src/main/java/jimmy/mcgymmy/logic/macro/NewMacroCommand.java
@@ -7,6 +7,7 @@ import jimmy.mcgymmy.logic.commands.CommandResult;
 import jimmy.mcgymmy.logic.commands.exceptions.CommandException;
 import jimmy.mcgymmy.logic.macro.exceptions.DuplicateMacroException;
 import jimmy.mcgymmy.logic.parser.exceptions.ParseException;
+import jimmy.mcgymmy.model.MacroList;
 import jimmy.mcgymmy.model.Model;
 
 public class NewMacroCommand implements CommandExecutable {
