+++ src/main/java/jimmy/mcgymmy/logic/parser/McGymmyParser.java
@@ -10,7 +10,7 @@ import org.apache.commons.cli.Options;
 import jimmy.mcgymmy.commons.core.Messages;
 import jimmy.mcgymmy.logic.commands.CommandExecutable;
 import jimmy.mcgymmy.logic.macro.Macro;
-import jimmy.mcgymmy.logic.macro.MacroList;
+import jimmy.mcgymmy.model.MacroList;
 import jimmy.mcgymmy.logic.macro.NewMacroCommand;
 import jimmy.mcgymmy.logic.parser.exceptions.ParseException;
