+++ src/test/java/jimmy/mcgymmy/logic/parser/McGymmyParserTest.java
@@ -11,7 +11,7 @@ import jimmy.mcgymmy.logic.commands.AddCommand;
 import jimmy.mcgymmy.logic.commands.CommandExecutable;
 import jimmy.mcgymmy.logic.commands.ListCommand;
 import jimmy.mcgymmy.logic.macro.Macro;
-import jimmy.mcgymmy.logic.macro.MacroList;
+import jimmy.mcgymmy.model.MacroList;
 import jimmy.mcgymmy.logic.macro.NewMacroCommand;
 import jimmy.mcgymmy.logic.parser.exceptions.ParseException;
