+++ src/test/java/jimmy/mcgymmy/storage/JsonMacroListStorageTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 
 import jimmy.mcgymmy.commons.exceptions.DataConversionException;
 import jimmy.mcgymmy.logic.macro.Macro;
-import jimmy.mcgymmy.logic.macro.MacroList;
+import jimmy.mcgymmy.model.MacroList;
 
 public class JsonMacroListStorageTest {
     private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonMacroListStorageTest");
